+++ cc/10_container_with_most_water.cc
@@ -2,6 +2,7 @@
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java:
 // https://pvs-studio.com
 #include <algorithm>
+#include <cstddef>
 #include <vector>
 
 #include "gtest/gtest.h"
@@ -11,9 +12,9 @@ namespace {
 class Solution {
    public:
     int MaxArea(std::vector<int>& height) {
-        int ret = 0;
-        int i = 0;
-        int j = height.size() - 1;
+        size_t ret = 0;
+        size_t i = 0;
+        size_t j = height.size() - 1;
         for (; j > i;) {
             int const het = std::min(height[i], height[j]);
             ret = std::max(ret, het * (j - i));
@@ -28,9 +29,9 @@ class Solution {
     }
 
     int MaxAreaV2(std::vector<int>& height) {
-        int l = 0;
-        int h = height.size() - 1;
-        int max_area = 0;
+        size_t l = 0;
+        size_t h = height.size() - 1;
+        size_t max_area = 0;
         while (l < h) {
             int const lh = height[l];
             int const hh = height[h];
